move test_arrays back to uint32_t for %a? array len

+++ tests/test_arrays_BE_lightweight.c
@@ -20,7 +20,7 @@ int main(int argc, char **argv)
     Packet_t* pkt = (Packet_t*)malloc(sizeof(Packet_t));
     Network_t * net;
     int tag=0;
-    uint64_t recv_array_len=0;
+    uint32_t recv_array_len=0;
     void * recv_array=NULL;
     int success=1;
     Rank my_rank;
